Move all test files to test folder
