+++ package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "auth0-lock": "^10.20.0",
     "jwt-decode": "^2.2.0",
+    "prop-types": "^15.5.10",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-redux": "^5.0.6",
