+++ src/containers/Header/Header.js
@@ -1,0 +1,21 @@
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
+
+import { authActions } from '../../redux/modules/auth';
+
+import HeaderView from './HeaderView';
+
+const mapStateToProps = state => ({
+  auth: state.auth,
+});
+
+const mapDispatchToProps = dispatch => ({
+  loginRequest: () => dispatch(authActions.loginRequest()),
+  logoutSuccess: () => dispatch(authActions.logoutSuccess()),
+});
+
+export default withRouter(connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(HeaderView));
+
