+++ test/unit/containers/App/AppView.test.js
