+++ test/unit/containers/Header/HeaderView.test.js
